Major rewrite for robustness / forgiveness
Allow named Geo sections, which seems to be a widely-used but
undocumented extension.

Allow using airport names as positions in other sections, which seems
to be occasionally-done in practice.

Be more forgiving in SID/STAR parsing: many files don't adhere to the
26 character formatting so instead try different chunkings until we
find a segment of 4 positions to infer where the name ends.

Improve robustness of lat/long parsing to allow partial specifications
like E40.22 and not require additional decimals.

Rework line handling to report accurate line numbers with errors.
(This had been broken)

Switch to string rather than []byte; the code is cleaner and we're
now unicode-friendly.

Switch to go-style error returns from parsing functions rather than
issuing syntax errors deep in the parse functions.

Be better about continuing onward (and not crashing) when things
are borked.