Optimize lat/long parsing
Previously we would take every lat-long and do a bunch of map
lookups to see if it was a VOR name (et al.) where we should use
the VOR's location. If that failed, then we'd try treating it as a
dotted latlong and parse that. Now, we look at what we've got
and try to figure out what it is ahead of time and call the dotted
value parser immediately when appropriate.

Bonus: got rid of an unnecessary atof that could have been an
atoi in the latlong parsing code for more speed. Also got rid of
some slow error formatting code.

For some sct2 files, this ends up making parsing 2.5-3x faster(!)